refactor: BridgeController에 Bridge와 관련없는 기능 분리
