+++ src/main/java/bridge/Application.java
@@ -1,5 +1,7 @@
 package bridge;
 
+import bridge.game.BridgeGame;
+
 public class Application {
 
     public static void main(String[] args) {
